Refactor:
changed the files structer, and edit the return statment scope
