+++ SimpleInventoryManagementSystem/models/IProductPrint.cs
@@ -1,0 +1,7 @@
+namespace SimpleInventoryManagementSystem.models;
+
+public interface IProductPrint
+{
+    void Print(List<Product> products);
+}
+
