+++ SimpleInventoryManagementSystem/models/IProductReader.cs
@@ -1,0 +1,6 @@
+namespace SimpleInventoryManagementSystem.models;
+
+public interface IProductReader
+{
+    List<Product> GetProducts();
+}
