+++ SimpleInventoryManagementSystem/models/IProductWrite.cs
@@ -1,0 +1,8 @@
+namespace SimpleInventoryManagementSystem.models;
+
+public interface IProductWriter
+{
+    bool AddProduct(Product product);
+    bool EditProduct(string oldProductName, Product updatedProduct);
+    bool DeleteProduct(string productName);
+}
