+++ SimpleInventoryManagementSystem/models/PrintProducts.cs
@@ -1,0 +1,12 @@
+namespace SimpleInventoryManagementSystem.models;
+
+public class PrintProducts : IProductPrint
+{
+    public void Print(List<Product> products)
+    {
+        foreach (var prod in products)
+        {
+            Console.WriteLine($"Product name: {prod.Name}, Quantity: {prod.Quantity}, Price: {prod.Price}");
+        }
+    }
+}
